avatar-agent/requirements.txt
livekit-agents>=0.12.0
livekit-plugins-simli>=0.2.0
livekit-plugins-google>=0.10.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
